To be compatible with other architectures use size_t everywhere instead of unsigned int

+++ fft/bitreverse.h
@@ -1,22 +1,24 @@
+#include <stddef.h>
+
 int gsl_fft_complex_bitreverse_order (gsl_complex data[], 
-				      const unsigned int n,
-				      const unsigned int logn) ;
+				      const size_t n,
+				      const size_t logn) ;
 
 int gsl_fft_real_bitreverse_order (double data[], 
-				   const unsigned int n,
-				   const unsigned int logn) ;
+				   const size_t n,
+				   const size_t logn) ;
 
 
 int gsl_fft_complex_goldrader_bitreverse_order (gsl_complex data[], 
-						const unsigned int n) ;
+						const size_t n) ;
 
 int gsl_fft_real_goldrader_bitreverse_order (double data[], 
-					     const unsigned int n) ;
+					     const size_t n) ;
 
 int gsl_fft_complex_rodriguez_bitreverse_order (gsl_complex data[], 
-						const unsigned int n,
-						const unsigned int logn) ;
+						const size_t n,
+						const size_t logn) ;
 
 int gsl_fft_real_rodriguez_bitreverse_order (double data[], 
-					     const unsigned int n,
-					     const unsigned int logn) ;
+					     const size_t n,
+					     const size_t logn) ;
