+++ fft/complex_init.c
@@ -8,11 +8,11 @@
 #include "factorize.h"
 
 int
-gsl_fft_complex_init (unsigned int n,
+gsl_fft_complex_init (size_t n,
 		      gsl_fft_complex_wavetable * wavetable)
 {
   int status;
-  unsigned int n_factors;
+  size_t n_factors;
 
   if (n == 0)
     {
@@ -41,12 +41,12 @@ gsl_fft_complex_init (unsigned int n,
 }
 
 int
-gsl_fft_complex_generate_wavetable (unsigned int n,
+gsl_fft_complex_generate_wavetable (size_t n,
 				    gsl_fft_complex_wavetable * wavetable)
 {
-  unsigned int i;
+  size_t i;
   double d_theta;
-  unsigned int t, product, product_1, q;
+  size_t t, product, product_1, q;
 
   if (n == 0)
     {
@@ -64,8 +64,8 @@ gsl_fft_complex_generate_wavetable (unsigned int n,
   product = 1;
   for (i = 0; i < wavetable->nf; i++)
     {
-      unsigned int j;
-      const unsigned int factor = wavetable->factor[i];
+      size_t j;
+      const size_t factor = wavetable->factor[i];
       wavetable->twiddle[i] = wavetable->trig + t;
       product_1 = product;	/* product_1 = p_(i-1) */
       product *= factor;
@@ -73,8 +73,8 @@ gsl_fft_complex_generate_wavetable (unsigned int n,
 
       for (j = 1; j < factor; j++)
 	{
-	  unsigned int k;
-	  unsigned int m = 0;
+	  size_t k;
+	  size_t m = 0;
 	  for (k = 1; k <= q; k++)
 	    {
 	      double theta;
@@ -98,7 +98,7 @@ gsl_fft_complex_generate_wavetable (unsigned int n,
 }
 
 int
-gsl_fft_complex_wavetable_alloc (unsigned int n,
+gsl_fft_complex_wavetable_alloc (size_t n,
 				 gsl_fft_complex_wavetable * wavetable)
 {
   if (n == 0)
