+++ fft/complex_pass_5.c
@@ -8,21 +8,21 @@ int
 gsl_fft_complex_pass_5 (const gsl_complex from[],
 			gsl_complex to[],
 			const gsl_fft_direction sign,
-			const unsigned int product,
-			const unsigned int n,
+			const size_t product,
+			const size_t n,
 			const gsl_complex twiddle1[],
 			const gsl_complex twiddle2[],
 			const gsl_complex twiddle3[],
 			const gsl_complex twiddle4[])
 {
-  unsigned int i = 0, j = 0;
-  unsigned int k, k1;
+  size_t i = 0, j = 0;
+  size_t k, k1;
 
-  const unsigned int factor = 5;
-  const unsigned int m = n / factor;
-  const unsigned int q = n / product;
-  const unsigned int product_1 = product / factor;
-  const unsigned int jump = (factor - 1) * product_1;
+  const size_t factor = 5;
+  const size_t m = n / factor;
+  const size_t q = n / product;
+  const size_t product_1 = product / factor;
+  const size_t jump = (factor - 1) * product_1;
 
   const double sin_2pi_by_5 = sin (2.0 * M_PI / 5.0);
   const double sin_2pi_by_10 = sin (2.0 * M_PI / 10.0);
@@ -80,11 +80,11 @@ gsl_fft_complex_pass_5 (const gsl_complex from[],
 	    x3_real, x3_imag, x4_real, x4_imag;
 
 	  {
-	    const unsigned int from0 = i;
-	    const unsigned int from1 = from0 + m;
-	    const unsigned int from2 = from1 + m;
-	    const unsigned int from3 = from2 + m;
-	    const unsigned int from4 = from3 + m;
+	    const size_t from0 = i;
+	    const size_t from1 = from0 + m;
+	    const size_t from2 = from1 + m;
+	    const size_t from3 = from2 + m;
+	    const size_t from4 = from3 + m;
 
 	    z0 = from[from0];
 	    z1 = from[from1];
@@ -166,11 +166,11 @@ gsl_fft_complex_pass_5 (const gsl_complex from[],
 
 	  /* apply twiddle factors */
 	  {
-	    const unsigned int to0 = j;
-	    const unsigned int to1 = to0 + product_1;
-	    const unsigned int to2 = to1 + product_1;
-	    const unsigned int to3 = to2 + product_1;
-	    const unsigned int to4 = to3 + product_1;
+	    const size_t to0 = j;
+	    const size_t to1 = to0 + product_1;
+	    const size_t to2 = to1 + product_1;
+	    const size_t to3 = to2 + product_1;
+	    const size_t to4 = to3 + product_1;
 
 	    /* to0 = 1 * x0 */
 	    to[to0].real = x0_real;
