+++ fft/fft_real.h
@@ -2,22 +2,22 @@
 int
   gsl_fft_real_pass_2 (const double from[],
 		       double to[],
-		       const unsigned int product,
-		       const unsigned int n,
+		       const size_t product,
+		       const size_t n,
 		       const gsl_complex twiddle[]);
 
 int
   gsl_fft_real_pass_3 (const double from[], double to[],
-		       const unsigned int product,
-		       const unsigned int n,
+		       const size_t product,
+		       const size_t n,
 		       const gsl_complex twiddle1[],
 		       const gsl_complex twiddle2[]);
 
 int
   gsl_fft_real_pass_4 (const double from[],
 		       double to[],
-		       const unsigned int product,
-		       const unsigned int n,
+		       const size_t product,
+		       const size_t n,
 		       const gsl_complex twiddle1[],
 		       const gsl_complex twiddle2[],
 		       const gsl_complex twiddle3[]);
@@ -25,8 +25,8 @@ int
 int
   gsl_fft_real_pass_5 (const double from[],
 		       double to[],
-		       const unsigned int product,
-		       const unsigned int n,
+		       const size_t product,
+		       const size_t n,
 		       const gsl_complex twiddle1[],
 		       const gsl_complex twiddle2[],
 		       const gsl_complex twiddle3[],
@@ -34,14 +34,14 @@ int
 
 int
   gsl_fft_real_pass_6 (double *from, double *to,
-		       unsigned int product, unsigned int n,
+		       size_t product, size_t n,
 		       gsl_complex * twiddle1, gsl_complex * twiddle2,
 		       gsl_complex * twiddle3, gsl_complex * twiddle4,
 		       gsl_complex * twiddle5);
 
 int
   gsl_fft_real_pass_n (const double from[], double to[],
-		       const unsigned int factor,
-		       const unsigned int product,
-		       const unsigned int n,
+		       const size_t factor,
+		       const size_t product,
+		       const size_t n,
 		       const gsl_complex twiddle[]);
